Refactor DraftlyWrapper to utilize custom hook for canvas event handling
* Simplified the DraftlyWrapper component by extracting pointer event handling logic into a custom hook, useCanvasWrapper. This enhances code clarity and maintainability while preserving existing functionality.
